Module s2n_quic_transport::connection

This module contains the implementation of QUIC Connections and their management

Modules

close

error

id

Defines the QUIC connection ID

limits

Structs

Connection

A QUIC connection

Implementation

InitialId

Uniquely identifies a QUIC connection between 2 peers

Limits

LocalId

Uniquely identifies a QUIC connection between 2 peers

OpenToken

An opaque token issued to each connection handle which allows the stream controller to track any pending open requests.

Parameters

Parameters which are passed to a Connection. These are unique per created connection.

PeerId

Uniquely identifies a QUIC connection between 2 peers

UnboundedId

Uniquely identifies a QUIC connection between 2 peers

Enums

Error

Errors that a connection can encounter.

ProcessingError

Some connection methods may need to indicate both TransportErrors and CryptoErrors. This enum is used to allow for either error type to be returned as appropriate.

Traits

Lock

A lock that synchronizes connection state between the QUIC endpoint thread and application