Struct s2n_quic_transport::connection::Connection

pub struct Connection { /* private fields */ }

A QUIC connection

Implementations

impl Connection

pub fn poll_accept(
&mut self,
stream_type: Option<StreamType>,
context: &Context<'_>
) -> Poll<Result<Option<Stream>, Error>>

Accepts an incoming Stream

The method will return

Poll::Ready(Ok(Some(stream, stream_type))) if a Stream was accepted

Poll::Ready(Ok(None)) if the connection was closed without an error

Poll::Ready(Err(stream_error)) if no stream could be accepted due to an error

Poll::Pending if no new Stream of the given type was accepted by the connection yet. In this case the caller must retry calling Self::poll_accept. For this purpose the method will save the core::task::Waker which is provided as part of the Context parameter, and notify it as soon as retrying the method will yield a different result.

pub fn poll_open_stream(
&mut self,
stream_type: StreamType,
context: &Context<'_>
) -> Poll<Result<Stream, Error>>

pub fn poll_request(
&self,
stream_id: StreamId,
request: &mut Request<'_>,
context: Option<&Context<'_>>
) -> Result<Response, StreamError>

pub fn close(&self, error_code: Error)

Closes the Connection with the provided error code

This will immediately terminate all outstanding streams.

pub fn server_name(&self) -> Result<Option<ServerName>, Error>

pub fn application_protocol(&self) -> Result<Bytes, Error>

pub fn id(&self) -> u64

pub fn ping(&self) -> Result<(), Error>

pub fn keep_alive(&self, enabled: bool) -> Result<(), Error>

pub fn local_address(&self) -> Result<SocketAddress, Error>

pub fn remote_address(&self) -> Result<SocketAddress, Error>

pub fn query_event_context(&self, query: &mut dyn Query) -> Result<(), Error>

pub fn query_event_context_mut(
&self,
query: &mut dyn QueryMut
) -> Result<(), Error>

pub fn datagram_mut(&self, query: &mut dyn QueryMut) -> Result<(), Error>