Struct s2n_quic_transport::endpoint::Context

pub struct Context<'a, Cfg: Config> {

    pub congestion_controller: &'a mut Cfg::CongestionControllerEndpoint,
    pub connection_id_format: &'a mut Cfg::ConnectionIdFormat,
    pub stateless_reset_token_generator: &'a mut Cfg::StatelessResetTokenGenerator,
    pub random_generator: &'a mut Cfg::RandomGenerator,
    pub tls: &'a mut Cfg::TLSEndpoint,
    pub endpoint_limits: &'a mut Cfg::EndpointLimits,
    pub token: &'a mut Cfg::TokenFormat,
    pub connection_limits: &'a mut Cfg::ConnectionLimits,
    pub connection_close_formatter: &'a mut Cfg::ConnectionCloseFormatter,
    pub event_subscriber: &'a mut Cfg::EventSubscriber,
    pub path_migration: &'a mut Cfg::PathMigrationValidator,
    pub packet_interceptor: &'a mut Cfg::PacketInterceptor,
    pub datagram: &'a mut Cfg::DatagramEndpoint,

}

Fields
congestion_controller: &'a mut Cfg::CongestionControllerEndpoint

The congestion controller endpoint associated with the endpoint config

connection_id_format: &'a mut Cfg::ConnectionIdFormat

The connection id format associated with the endpoint config

stateless_reset_token_generator: &'a mut Cfg::StatelessResetTokenGenerator

The stateless reset token generator associated with the endpoint config

random_generator: &'a mut Cfg::RandomGenerator

The random data generator associated with the endpoint config

tls: &'a mut Cfg::TLSEndpoint

The TLS endpoint associated with the endpoint config

endpoint_limits: &'a mut Cfg::EndpointLimits

The endpoint limits

token: &'a mut Cfg::TokenFormat

Token generator / validator

connection_limits: &'a mut Cfg::ConnectionLimits

The connection limits

connection_close_formatter: &'a mut Cfg::ConnectionCloseFormatterevent_subscriber: &'a mut Cfg::EventSubscriberpath_migration: &'a mut Cfg::PathMigrationValidatorpacket_interceptor: &'a mut Cfg::PacketInterceptordatagram: &'a mut Cfg::DatagramEndpoint