Struct s2n_quic_transport::recovery::Manager

pub struct Manager<Config: Config> { /* private fields */ }

Implementations

impl<Config: Config> Manager<Config>

pub fn new(space: PacketNumberSpace) -> Self

Constructs a new recovery::Manager

pub fn on_retry_packet<Pub: ConnectionPublisher>(
&mut self,
path: &mut Path<Config>,
path_id: Id,
publisher: &mut Pub
)

Invoked when the Client processes a Retry packet.

Reset congestion controller state by discarding sent bytes and replacing recovery manager with a new instance of itself.

pub fn on_timeout<Ctx: Context<Config>, Pub: ConnectionPublisher>(
&mut self,
timestamp: Timestamp,
random_generator: &mut Config::RandomGenerator,
context: &mut Ctx,
publisher: &mut Pub
)

pub fn on_packet_sent<Ctx: Context<Config>, Pub: ConnectionPublisher>(
&mut self,
packet_number: PacketNumber,
outcome: Outcome,
time_sent: Timestamp,
ecn: ExplicitCongestionNotification,
transmission_mode: Mode,
app_limited: Option<bool>,
context: &mut Ctx,
publisher: &mut Pub
)

pub fn update_pto_timer(
&mut self,
path: &Path<Config>,
now: Timestamp,
is_handshake_confirmed: bool
)

Updates the PTO timer

pub fn on_transmit<W: WriteContext>(&mut self, context: &mut W)

Queries the component for any outgoing frames that need to get sent

pub fn on_ack_frame<A: AckRanges, Ctx: Context<Config>, Pub: ConnectionPublisher>(
&mut self,
timestamp: Timestamp,
frame: Ack<A>,
packet_number: PacketNumber,
random_generator: &mut Config::RandomGenerator,
context: &mut Ctx,
publisher: &mut Pub
) -> Result<(), Error>

Process ACK frame.

Update congestion controller, timers and meta data around acked packet ranges.

pub fn requires_probe(&self) -> bool

Returns true if the recovery manager requires a probe packet to be sent.

pub fn on_packet_number_space_discarded<Pub: ConnectionPublisher>(
&mut self,
path: &mut Path<Config>,
path_id: Id,
publisher: &mut Pub
)

Clears bytes in flight for sent packets.