Struct s2n_quic_transport::stream::Controller

pub struct Controller { /* private fields */ }

This component manages stream concurrency limits.

It enforces both the local initiated stream limits and the peer initiated stream limits.

It will also signal an increased max streams once streams have been consumed.

Implementations

impl Controller

pub fn new(
local_endpoint_type: Type,
initial_peer_limits: InitialFlowControlLimits,
initial_local_limits: InitialFlowControlLimits,
stream_limits: Limits
) -> Self

Creates a new stream::Controller

The peer will be allowed to open streams up to the given initial_local_limits.

For local_initiated unidirectional streams, the local application will be allowed to open up to the minimum of the given local limits (stream_limits) and initial_peer_limits.

For bidirectional streams, the local application will be allowed to open up to the minimum of the given initial_local_limits and initial_peer_limits.

The peer may give additional credit to open more streams by delivering MAX_STREAMS frames.

pub fn on_max_streams(&mut self, frame: &MaxStreams)

This method is called when a MAX_STREAMS frame is received, which signals an increase in the available streams budget.

pub fn poll_open_local_stream(
&mut self,
stream_type: StreamType,
open_tokens: &mut OpenToken,
context: &Context<'_>
) -> Poll<()>

This method is called when the local application wishes to open the next stream of a type (Bidirectional/Unidirectional).

Poll::Pending is returned when there isn’t available capacity to open a stream, either because of local initiated concurrency limits or the peer’s stream limits. If Poll::Pending is returned, the waker in the given context will be woken when additional stream capacity becomes available.

pub fn on_open_remote_stream(
&mut self,
stream_iter: StreamIter
) -> Result<(), Error>

This method is called when the remote peer wishes to open a new stream.

Opening a Stream also opens all lower Streams of the same type. Therefore this function validates if there is enough capacity to open all streams.

A STREAM_LIMIT_ERROR will be returned if the peer has exceeded the stream limits that were communicated by transport parameters or MAX_STREAMS frames.

pub fn on_close_stream(&mut self, stream_id: StreamId)

This method is called whenever a stream is closed.

pub fn close(&mut self)

This method is called when the stream manager is closed. All wakers will be woken to unblock waiting tasks.

pub fn on_packet_ack<A: Set>(&mut self, ack_set: &A)

This method is called when a packet delivery got acknowledged

pub fn on_packet_loss<A: Set>(&mut self, ack_set: &A)

This method is called when a packet loss is reported

pub fn update_blocked_sync_period(&mut self, blocked_sync_period: Duration)

Updates the period at which STREAMS_BLOCKED frames are sent to the peer if the application is blocked by peer limits.

pub fn on_transmit<W: WriteContext>(
&mut self,
context: &mut W
) -> Result<(), OnTransmitError>

Queries the component for any local_initiated frames that need to get sent

pub fn on_timeout(&mut self, now: Timestamp)

Called when the connection timer expires